// src/interfaces/categorias.interfaces.ts
// Generated by https://quicktype.io

import { IUser } from "./auth.interfaces";
import { IContacto } from "./contactos.interfaces";

export interface ICategoria {
    id:         number;
    attributes: ICategoriaAttributes;
}

export interface ICategoriaAttributes {
    nombre:      string;
    createdAt:   string;
    updatedAt:   string;
    publishedAt: string;
    contactos:   {
        data: IContacto[]
    };
    usuario:  {
        data: IUser
    }
}

export interface Contactos {
    data: any[];
}